Simplify node dirty tile script to use SQL function
